Ensure thatPreferredCursorExecution is the only configuration flag to control cursor preference.
[resolves #267]
