+++ src/test/java/io/r2dbc/mssql/SimpleMssqlStatementUnitTests.java
@@ -24,16 +24,7 @@ import io.r2dbc.mssql.message.Message;
 import io.r2dbc.mssql.message.TransactionDescriptor;
 import io.r2dbc.mssql.message.tds.Encode;
 import io.r2dbc.mssql.message.tds.ServerCharset;
-import io.r2dbc.mssql.message.token.Column;
-import io.r2dbc.mssql.message.token.ColumnMetadataToken;
-import io.r2dbc.mssql.message.token.DataToken;
-import io.r2dbc.mssql.message.token.DoneToken;
-import io.r2dbc.mssql.message.token.ErrorToken;
-import io.r2dbc.mssql.message.token.RowToken;
-import io.r2dbc.mssql.message.token.RowTokenFactory;
-import io.r2dbc.mssql.message.token.RpcRequest;
-import io.r2dbc.mssql.message.token.SqlBatch;
-import io.r2dbc.mssql.message.token.Tabular;
+import io.r2dbc.mssql.message.token.*;
 import io.r2dbc.mssql.message.type.Collation;
 import io.r2dbc.mssql.message.type.LengthStrategy;
 import io.r2dbc.mssql.message.type.SqlServerType;
@@ -42,8 +33,6 @@ import io.r2dbc.spi.ColumnMetadata;
 import io.r2dbc.spi.R2dbcNonTransientResourceException;
 import io.r2dbc.spi.Result;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ValueSource;
 import org.mockito.ArgumentCaptor;
 import org.reactivestreams.Publisher;
 import reactor.core.publisher.Flux;
@@ -61,9 +50,7 @@ import java.util.function.Predicate;
 import static io.r2dbc.mssql.message.type.TypeInformation.Builder;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
+import static org.mockito.Mockito.*;
 
 /**
  * Unit tests for {@link SimpleMssqlStatement}.
@@ -79,7 +66,7 @@ class SimpleMssqlStatementUnitTests {
 
         createColumn(3, "salary", SqlServerType.MONEY, 8, LengthStrategy.BYTELENTYPE, null)).toArray(new Column[0]);
 
-    static final ConnectionOptions OPTIONS = new ConnectionOptions();
+    static final ConnectionOptions OPTIONS = new TestConnectionOptions();
 
     @Test
     void shouldReportNumberOfAffectedRows() {
@@ -352,15 +339,4 @@ class SimpleMssqlStatementUnitTests {
         return new Column(index, name, type, null);
     }
 
-    @ParameterizedTest
-    @ValueSource(strings = {"select", "SELECT", "sElEcT"})
-    void shouldAcceptQueries(String query) {
-        assertThat(SimpleMssqlStatement.prefersCursors(query)).isTrue();
-    }
-
-    @ParameterizedTest
-    @ValueSource(strings = {" select", "sp_cursor", "INSERT"})
-    void shouldRejectQueries(String query) {
-        assertThat(SimpleMssqlStatement.prefersCursors(query)).isFalse();
-    }
 }
